feat: check if user is owner of a product

+++ service/src/main/java/ar/edu/itba/paw/service/SellOrderServiceImpl.java
@@ -7,16 +7,19 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.math.BigDecimal;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
 public class SellOrderServiceImpl implements SellOrderService {
 
     private final SellOrderDao sellOrderDao;
+    private final UserService userService;
 
     @Autowired
-    public SellOrderServiceImpl(SellOrderDao sellOrderDao) {
+    public SellOrderServiceImpl(SellOrderDao sellOrderDao, UserService userService) {
         this.sellOrderDao = sellOrderDao;
+        this.userService = userService;
     }
 
     @Override
@@ -31,13 +34,23 @@ public class SellOrderServiceImpl implements SellOrderService {
 
     @Override
     public boolean update(long id, String category, BigDecimal price, String description) {
-        // TODO: check if current user is owner
+        if (!isUserOwner(id))
+            return false;
         return sellOrderDao.update(id, category, price, description);
     }
 
     @Override
     public boolean delete(long id) {
-        // TODO: check if current user is owner
+        if (!isUserOwner(id))
+            return false;
         return sellOrderDao.delete(id);
     }
+
+    @Override
+    public boolean isUserOwner(long sellOrderId) {
+        Optional<SellOrder> maybeOrder = getOrderById(sellOrderId);
+        String currentUserEmail = userService.getCurrentUser().getEmail();
+        return maybeOrder.isPresent() && Objects.equals(currentUserEmail, maybeOrder.get().getSellerEmail());
+    }
+
 }
